# Doc/swagger_subastar.yaml
openapi: 3.0.3
info:
  title: SubastAR API
  description: >
    API REST de SubastAR — app de subastas dinámicas ascendentes (TPO DA1, 1C2026).
    Cubre el dominio descripto en TPO_DAI_1C2026.pdf: registro en dos etapas con
    foto de DNI, gestión de medios de pago, subastas con categoría y moneda,
    pujas con reglas de mínimo/máximo, consignaciones con inspección y póliza
    de seguro, notificaciones, métricas y ventas.
  version: 2.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local
  - url: https://subastar.example.com/api/v1
    description: Producción (entrega 3)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [documento, password]
      properties:
        documento:
          type: string
          example: "30123456"
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        user: { $ref: '#/components/schemas/UserSummary' }
    UserSummary:
      type: object
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string }
        apellido: { type: string }
        email: { type: string, format: email }
        documento: { type: string }
        categoria: { $ref: '#/components/schemas/Categoria' }
        estado: { $ref: '#/components/schemas/EstadoUsuario' }
    Categoria:
      type: string
      enum: [COMUN, ESPECIAL, PLATA, ORO, PLATINO]
    EstadoUsuario:
      type: string
      enum: [PENDIENTE_VERIFICACION, APROBADO, SUSPENDIDO]
    RegisterStep1Request:
      type: object
      required: [documento, email, nombre, apellido, domicilioLegal, paisOrigen]
      properties:
        documento: { type: string }
        email: { type: string, format: email }
        nombre: { type: string }
        apellido: { type: string }
        domicilioLegal: { type: string }
        paisOrigen: { type: string, description: "ISO 3166-1 alpha-3", example: "ARG" }
        fotoDniFrente: { type: string, format: byte }
        fotoDniDorso: { type: string, format: byte }
    RegisterStep2Request:
      type: object
      required: [registrationToken, password, passwordConfirm]
      properties:
        registrationToken: { type: string }
        password: { type: string, format: password }
        passwordConfirm: { type: string, format: password }
...
    MedioPago:
      type: object
      properties:
        id: { type: string }
        tipo:
          type: string
          enum: [CUENTA_BANCARIA, TARJETA_CREDITO, CHEQUE_CERTIFICADO]
        proveedor: { type: string, example: "Mercado Pago / MasterCard / Banco Nación" }
        ultimosDigitos: { type: string, maxLength: 4 }
        verificado: { type: boolean }
        montoGarantia:
          type: number
          format: double
          nullable: true
          description: "Sólo aplica a CHEQUE_CERTIFICADO."
    AddPaymentMethodRequest:
      type: object
      required: [tipo, proveedor]
      properties:
        tipo: { type: string, enum: [CUENTA_BANCARIA, TARJETA_CREDITO, CHEQUE_CERTIFICADO] }
        proveedor: { type: string }
        ultimosDigitos: { type: string, maxLength: 4 }
        montoGarantia: { type: number, format: double, nullable: true }
    Consignment:
      type: object
      description: "Solicitud para subastar un artículo propio del usuario."
      properties:
        id: { type: string }
        nombreBien: { type: string }
        descripcionDetallada: { type: string }
        historia: { type: string, nullable: true }
        fotos:
          type: array
          items: { type: string }
          minItems: 6
        declaraPropiedad: { type: boolean }
        declaraOrigenLicito: { type: boolean }
        estado:
          type: string
          enum: [ENVIADA, EN_INSPECCION, ACEPTADA, RECHAZADA, EN_SUBASTA, VENDIDA]
        # Datos completados por la empresa cuando inspecciona y propone
        valorBaseOfrecido: { type: number, format: double, nullable: true }
        comision: { type: number, format: double, nullable: true }
        polizaSeguro:
          $ref: '#/components/schemas/PolizaSeguro'
          nullable: true
        ubicacionDeposito: { type: string, nullable: true }
        # Si es ACEPTADA, datos de la subasta asignada
        subastaAsignadaId: { type: string, nullable: true }
        fechaSubastaAsignada: { type: string, format: date-time, nullable: true }
        # Si es RECHAZADA
        motivoRechazo: { type: string, nullable: true }
        gastosDevolucion: { type: number, format: double, nullable: true }
    PolizaSeguro:
      type: object
      properties:
        id: { type: string }
        aseguradora: { type: string }
        numeroPoliza: { type: string }
        valorAsegurado: { type: number, format: double }
        contactoAseguradora: { type: string }
    CreateConsignmentRequest:
      type: object
      required: [nombreBien, descripcionDetallada, fotos, declaraPropiedad, declaraOrigenLicito]
      properties:
        nombreBien: { type: string }
        descripcionDetallada: { type: string }
        historia: { type: string }
        fotos:
          type: array
          items: { type: string, format: byte }
          minItems: 6
        declaraPropiedad: { type: boolean }
        declaraOrigenLicito: { type: boolean }
...
security:
  - bearerAuth: []
paths:
...
  /auth/register/step1:
    post:
      tags: [Auth]
      summary: Etapa 1 — datos personales + foto DNI
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterStep1Request' }
      responses:
        '201': { description: Registro iniciado. Se envía mail al usuario al aprobarse. }
        '400': { description: Datos inválidos. }
        '409': { description: Email o documento ya registrado. }
  /auth/register/step2:
    post:
      tags: [Auth]
      summary: Etapa 2 — setear contraseña con token de verificación
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterStep2Request' }
      responses:
        '200':
          description: Contraseña creada. Devuelve tokens.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '400': { description: Datos inválidos. }
        '410': { description: Token expirado. }
  /auth/login:
    post:
      tags: [Auth]
      summary: Login por documento + contraseña
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Login exitoso.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401': { description: Credenciales incorrectas. }
        '403': { description: Usuario suspendido o pendiente de aprobación. }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refrescar access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: Nuevo access token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
        '401': { description: Refresh token inválido o expirado. }

  # ─── USERS ─────────────────────────────────────────────────────────
...
  /payment-methods:
    get:
      tags: [PaymentMethods]
      summary: Listar medios de pago del usuario
      responses:
        '200':
          description: Lista.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MedioPago' }
    post:
      tags: [PaymentMethods]
      summary: Agregar un medio de pago
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddPaymentMethodRequest' }
      responses:
        '201':
          description: Creado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MedioPago' }
  /payment-methods/{id}:
    delete:
      tags: [PaymentMethods]
      summary: Eliminar un medio de pago
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Eliminado. }
        '404': { description: No encontrado. }

  # ─── CONSIGNMENTS ──────────────────────────────────────────────────
  /consignments:
    get:
      tags: [Consignments]
      summary: Mis consignaciones (pantalla Mis Subastas)
      responses:
        '200':
          description: Lista.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Consignment' }
    post:
      tags: [Consignments]
      summary: Solicitar subasta de un artículo propio (Subastar algo propio)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateConsignmentRequest' }
      responses:
        '201':
          description: Solicitud creada (estado ENVIADA).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Consignment' }
        '400':
          description: Datos inválidos (faltan declaraciones, menos de 6 fotos, etc).
  /consignments/{id}:
    get:
      tags: [Consignments]
      summary: Detalle de una consignación
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Detalle.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Consignment' }
        '404': { description: No encontrada. }
  /consignments/{id}/accept-offer:
    patch:
      tags: [Consignments]
      summary: Aceptar la propuesta (valor base + comisiones) — pantalla Solicitud Aceptada
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Aceptado. La pieza pasa al catálogo de la subasta asignada.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Consignment' }
  /consignments/{id}/reject-offer:
    patch:
      tags: [Consignments]
      summary: Rechazar la propuesta (se devuelve el bien con cargo)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                motivo: { type: string }
      responses:
        '200':
          description: Rechazado. Estado RECHAZADA.

  # ─── NOTIFICATIONS ─────────────────────────────────────────────────
...